+++ insim/src/net/blocking_impl/framed.rs
@@ -1,33 +1,17 @@
-use std::{fmt::Debug, time::Duration};
+use std::{
+    fmt::Debug,
+    io::{Read, Write},
+};
 
 use bytes::{BufMut, BytesMut};
-use if_chain::if_chain;
-use tokio::{
-    io::{AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt},
-    time::{self, timeout},
-};
 
-/// Read Write super trait
-pub trait AsyncReadWrite: AsyncRead + AsyncWrite + Debug + Unpin + Send + Sync {}
+use super::ReadWrite;
+use crate::{insim::Isi, net::Codec, result::Result, Error, Packet, DEFAULT_BUFFER_CAPACITY};
 
-impl<T: AsyncRead + AsyncWrite + Debug + Unpin + Send + Sync> AsyncReadWrite for T {}
-
-// #[cfg(feature = "websocket")]
-// use super::websocket::TungsteniteWebSocket;
-use crate::{
-    error::Error,
-    insim::Isi,
-    net::{Codec, DEFAULT_TIMEOUT_SECS},
-    packet::Packet,
-    result::Result,
-    DEFAULT_BUFFER_CAPACITY,
-};
-
-/// A unified wrapper around anything that implements [AsyncReadWrite].
-/// You probably really want to look at [Framed].
+/// A unified wrapper around anything that implements Read + Write.
 #[derive(Debug)]
 pub struct Framed {
-    inner: Box<dyn AsyncReadWrite>,
+    inner: Box<dyn ReadWrite>,
     codec: Codec,
     buffer: BytesMut,
     verify_version: bool,
@@ -35,7 +19,7 @@ pub struct Framed {
 
 impl Framed {
     /// Create a new FramedInner, which wraps some kind of network transport.
-    pub fn new(inner: Box<dyn AsyncReadWrite>, codec: Codec) -> Self {
+    pub fn new(inner: Box<dyn ReadWrite>, codec: Codec) -> Self {
         let buffer = BytesMut::with_capacity(DEFAULT_BUFFER_CAPACITY);
 
         Self {
@@ -54,14 +38,14 @@ impl Framed {
     /// Performs the Insim handshake by sending a [Isi] packet.
     /// If the handshake does not complete within the given timeout, it will fail and the
     /// connection should be considered invalid.
-    pub async fn handshake(&mut self, isi: Isi, timeout: Duration) -> Result<()> {
-        time::timeout(timeout, self.write(isi)).await??;
+    pub fn handshake(&mut self, isi: Isi) -> Result<()> {
+        self.write(Into::<Packet>::into(isi))?;
 
         Ok(())
     }
 
     #[allow(unsafe_code)]
-    async fn read_buf(&mut self) -> Result<usize> {
+    fn read_buf(&mut self) -> Result<usize> {
         // We're making the assumption that the internal stream will not discard data if there is
         // not enough space in the chunk.
         // This is how TcpStream works, and how we've made UdpStream work.
@@ -72,7 +56,7 @@ impl Framed {
             let ptr = chunk.as_mut_ptr();
             // SAFETY: The inner stream is not going to read from the uninitialized data.
             let slice = unsafe { std::slice::from_raw_parts_mut(ptr, len) };
-            self.inner.read(slice).await?
+            self.inner.read(slice)?
         };
 
         // SAFETY: The inner stream just initialized that many bytes, we need to trust that
@@ -83,10 +67,10 @@ impl Framed {
         Ok(size)
     }
 
-    /// Asynchronously wait for a packet from the inner network.
-    pub async fn read(&mut self) -> Result<Packet> {
+    /// Wait for a packet from the inner network.
+    pub fn read(&mut self) -> Result<Packet> {
         loop {
-            if_chain! {
+            if_chain::if_chain! {
                 if !self.buffer.is_empty();
                 if let Some(packet) = self.codec.decode(&mut self.buffer)?;
                 then {
@@ -98,14 +82,14 @@ impl Framed {
                     // keepalive
                     if let Some(pong) = packet.maybe_pong() {
                         tracing::debug!("Ping? Pong!");
-                        self.write(pong).await?;
+                        self.write(pong)?;
                     }
 
                     return Ok(packet);
                 }
             }
 
-            match timeout(Duration::from_secs(DEFAULT_TIMEOUT_SECS), self.read_buf()).await? {
+            match self.read_buf() {
                 Ok(0) => {
                     // The remote closed the connection. For this to be a clean
                     // shutdown, there should be no data in the read buffer. If
@@ -124,17 +108,18 @@ impl Framed {
                     continue;
                 },
                 Err(e) => {
+                    tracing::info!("did get err={:?}", e);
                     return Err(e);
                 },
             }
         }
     }
 
-    /// Asynchronously write a packet to the inner network.
-    pub async fn write<P: Into<Packet>>(&mut self, packet: P) -> Result<()> {
-        let mut buf = self.codec.encode(&packet.into())?;
+    /// Write a packet to the inner network.
+    pub fn write<P: Into<Packet>>(&mut self, packet: P) -> Result<()> {
+        let buf = self.codec.encode(&packet.into())?;
         if !buf.is_empty() {
-            self.inner.write_all_buf(&mut buf).await?;
+            let _ = self.inner.write(&buf)?;
         }
 
         Ok(())
