+++ insim/src/net/tokio_impl/mod.rs
@@ -1,6 +1,4 @@
-// pub(crate) mod bufwriter;
 pub(crate) mod framed;
-// pub(crate) mod tcp;
 pub mod udp;
 
 #[cfg(feature = "websocket")]
